dvi_hdmi: silence build warnings in vendored code
common_dvi_pin_configs.h: every dvi_serialiser_cfg literal leaves
prog_offs (pure output state, written by dvi_serialiser_init() itself)
at its implicit zero, tripping -Wmissing-field-initializers on 9 structs.
Wrapped in a diagnostic push/ignore/pop rather than adding a redundant
`.prog_offs = 0` to each -- keeps upstream's own struct literals
untouched.

tmds_encode.c: tmds_setup_palette_symbols()/tmds_setup_palette24_symbols()
looped with `int i` against a `size_t n_palette` bound (-Wsign-compare);
changed the loop variable to size_t, matching the parameter it's compared
against -- both are only ever called with n_palette a small power of 2
(<=256 per their own doc comments), so this is a pure type-correctness
fix with no behavioral change.

Found via a from-scratch rebuild of every profile after the DVI/HDMI
migration + USB HID fixes landed; build-verified on pizero_lcd and
pizero_hdmi (0 warnings from this component on either).

+++ components/dvi_hdmi/tmds_encode.c
@@ -278,7 +278,7 @@ void tmds_setup_palette_symbols(const uint16_t *palette, uint32_t *tmds_palette,
 	uint32_t* tmds_palette_blue = tmds_palette;
 	uint32_t* tmds_palette_green = tmds_palette + 2 * n_palette;
 	uint32_t* tmds_palette_red = tmds_palette + 4 * n_palette;
-	for (int i = 0; i < n_palette; ++i) {
+	for (size_t i = 0; i < n_palette; ++i) {
 		uint16_t blue = (palette[i] << 3) & 0xf8;
 		uint16_t green = (palette[i] >> 3) & 0xfc;
 		uint16_t red = (palette[i] >> 8) & 0xf8;
@@ -296,7 +296,7 @@ void tmds_setup_palette24_symbols(const uint32_t *palette, uint32_t *tmds_palett
 	uint32_t* tmds_palette_blue = tmds_palette;
 	uint32_t* tmds_palette_green = tmds_palette + 2 * n_palette;
 	uint32_t* tmds_palette_red = tmds_palette + 4 * n_palette;
-	for (int i = 0; i < n_palette; ++i) {
+	for (size_t i = 0; i < n_palette; ++i) {
 		uint16_t blue = palette[i] & 0xff;
 		uint16_t green = (palette[i] >> 8) & 0xff;
 		uint16_t red = (palette[i] >> 16) & 0xff;
